requirements.txt
requests>=2.31.0
Pillow>=10.0.0
Flask>=3.0.0
gunicorn>=22.0.0
